test: 테스트 분리
api 단위로 테스트 분리
