+++ src/main/java/roomescape/RoomescapeApplication.java
@@ -5,7 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class RoomescapeApplication {
-
+    
     public static void main(String[] args) {
         SpringApplication.run(RoomescapeApplication.class, args);
     }
